I moved the constants from the project to a separate folder. I didn't make any changes, the status can be viewed by the previous commit.
