+++ src/service/repeat_message.py
@@ -10,7 +10,7 @@ from datetime import timedelta
 from telegram import InlineKeyboardButton, Update
 from telegram.ext import CallbackContext, ContextTypes
 
-from src.constants.command_constants import REPEAT_COMMAND
+from constants.command_constants import REPEAT_COMMAND
 
 
 async def repeat_message_after_1_hour_callback(update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
